Increment release version for next release
We will follow a standard MAJOR.MINOR.PATCH scheme, where the major
version is incremented when a change is backwards-incompatible.

The reason for incrementing the version now is so that if a development
version of Seastar is installed in a system, then it will have a higher
version number than a released version of Seastar.

Once a release is actually made, the major version increment can be
changed to a minor one (depending on the scope of changes since the last
release).