tests: skip the alien test if smp::count < 2
currently alien poller lives in smp_pollfn, which is not registered if
smp::count < 2. strictly speaking alien poller does not necessarily
relies on smp_pollfn, and we can registered them separately. but it's
simpler and probably faster if we keep current behavior, because we
can save a recurring task for reactor.

in this change, if smp::count is less than 2, no tasks is submitted from
alien, and a message is printed instead noting that the test is skipped.

Fixes #499
Signed-off-by: Kefu Chai <kchai@redhat.com>