build: make Sanitizers optional
as there is a chance that we need to stick with an ancient linker which
will fail to link with "-fsanitize=undefined" option:

/usr/bin/ld: unrecognized option '--push-state--no-as-needed'
/usr/bin/ld: use the --help option for usage information
collect2: error: ld returned 1 exit status

for instance, the `ld` from binutils on ubuntu 16.04 fails to work.

because Sanitizer_FOUND is set if both `-fsanitize=address` and
`-fsanitize=undefined` are supported. as `Sanitizers_FIBER_SUPPORT`
is checked separately, and does not depends on the availability of
address sanitizer and/or UB sanitizer, we can still enable it
as long as `Sanitizers_FIBER_SUPPORT` is set.

Fixes #562

Signed-off-by: Kefu Chai <kchai@redhat.com>