resource: set coordinator_to_idx if !SEASTAR_HAVE_HWLOC
if hwloc is not available we need to set coordinator_to_idx as well,
even it is known that it's always a shard always uses itself as the
coordinator and that `coordinator_to_idx[n] == n`. so we can populate
engine's ioqueue in the same way with or without HWLOC, instead of
adding `#ifdef SEASTAR_HAVE_HWLOC` smp::configure().

without this change, smp::configure() segfaults if Seastar_HWLOC is OFF.

Signed-off-by: Kefu Chai <kchai@redhat.com>