doc: update split script to use xml.etree.ElementTree
the htmlsplit.py failed to split the HTML generated with pandoc 2.9.2.1.
because of the change in the HTML schema.

instead of using regex to match the output, use xml.etree.ElementTree to
parse and modify the XML tree in hope that it'd be simpler to adapt to
future changes in the output schema.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>