cmake: check for DPDK only if enabled
so the configure.py is less chatty. and it's less confusing if
SPDK is enabled when configuration Seastar. because SPDK comes
with a bundled DPDK, whose version is not necessarily identical to
the one in the top-level dpdk submodule. so the output of cmake
when checking DPDK does not imply the existence of the DPDK used
by SPDK. since the DPDK and SPDK supports cannot be enabled at
the same time, it'd be less confusing if DPDK is checked only if
it is enabled.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>