memory: always allocate buf using "malloc" for non reactor
before this change, there is chance that Seastar reallocate a memory
buffer allocated by the c runtime even before "original_malloc_func"
is initialized. the runtime might want to allocate memory for some TLS
variables or just for itself, like env variables.

but we might want to call functions like setenv() from a reactor thread.
and setenv() could reallocates the memory buffer previously allocated by
the "main" thread. in this case, cpu_pages::shrink() panics if it is
asked to resize a memory chunk allocated by another reactor thread.

in this change, if allocate() is called in a non-reactor thread, the
original malloc() is always used. so it's not allocated in the Seastar
memory pool, hence the assertion failure can be avoided.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>