tests: unix_domain_test: fix thread/future confusion in client_round()
client_round() is partially written as a function returning a future
and partially as a function running in a thread. No harm was observed,
but it's asking for trouble.

Since it happens to run in a thread (and is waited for immediately),
make all of it be a threaded function.