log: add colored logging support
in order to improve visual debugging experience, logging message
with colored prefix could be helpful.

in this change, an option named "Seastar_COLORED_LOG" is added.
to minimize the build-time requirement and suprise to developers/users,
this option is disabled by default.

this change utilizes the color output from fmtlib. the colored
printing support was introduced to fmtlib 5.2.1 in the sense we
can find it in fmt/color.h. but the format helpers was added in
fmtlib 6.0.0. so the minimal fmt version required for enabling
"Seastar_COLORED_LOG" is 6.0.0.

because, apparently, this features is implemented using ANSI
escape code, this adds the amount of text printed to the specified
output, so if this extra overhead is a concern, it might be
advisable to disable this feature in production.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>