reactor_backend, io_queue: accept `const io_request&`
instead of using `io_request&` as the parameter type, using `const
io_request&` instead. as we don't mutate the parameter at the callee
side. we just read interpret the request. so `const io_request&` should
be a better option in this use case. and it sends a more accurate
message to the human readers.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>