build: cooking.sh: parse -s "foo=bar=baz" as foo="bar=bz"
'-s VAR=VALUE' is used to set an environmental variable of
"VAR=VALUE" when calling CMake. but if we want to set a CFLAGS like
"CFLAGS=-stdlib=libc++", the existing parse_assignment()

1. splits the string into three parts and
2. assigns it to an array,
3. uses the 2nd element in the array as the value of the environmental
   variable

in the use case above, before this change, what we expect is an
environmental variable with value of "-stdlib=libc++", but what we get
is "-stdlib".

after this change, the part after the first delimiter is considered
as the value of the environmental variable.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>