build: customize toolset for building Boost
before this change, CMAKE_C_FLAGS and CMAKE_CXX_FLAGS are not populated
down to recipes which should also be compiled with the same set of
flags, especially when it comes to ingredients implemented using C++,
where the compiling flags might be critical as they could be used to
pick the standard library and linker flags. and the built libraries
built with different set of compiling flags could be incompatible with
each other. so we have to use the same set of flags when building
C++ projects.

also, we should use the same C++ compiler for building C++ project
and for building Seastar. as different C++ compiler might accept
different sets of options.

after this change, cook_boost.jam is used to customize the toolset used
for building buliding Boost. in this file, both the compiler and the
compiling options are set according to the cmake settings.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>