build: bump fmtlib to 9.1.0 in cooking_recipe
in hope to help developers to verify their change on an older distro
where the version of stock fmtlib is lower than 9.0.0.

because, fmtlib v9 has a couple of breaking changes. there are chances
a change builds locally with an older version of fmtlib but it just
fails the build of fmtlib v9. this change would help to alleviate
this issue, if the dependencies are prepared using cooking.sh .

Signed-off-by: Kefu Chai <tchaikov@gmail.com>