build: make the header ragel-generated target SYSTEM
as ragel could generate headers which does not meet the criteria
enforced by the project, so the compiler may warn and fail the
build like:

```
/home/runner/work/scylladb/scylladb/build/gen/redis/protocol_parser.hh:110:1: error: unannotated fall-through between switch labels [clang-diagnostic-implicit-fallthrough]
  110 | case 1:
      | ^
/home/runner/work/scylladb/scylladb/build/gen/redis/protocol_parser.hh:110:1: note: insert 'FMT_FALLTHROUGH;' to silence this warning
  110 | case 1:
      | ^
      | FMT_FALLTHROUGH;
```

but the seastar application cannot control the behavior of ragel.
if it is a CMake-based project, what it can do is to mark the headers
as SYSTEM headers.

both Seastar and Seastar applications suffer from this. but Seastar can
address by marking the exposed target with SYSTEM header. if the Seastar
project chooses to include Seastar by including its CMakeLists.txt and
use `seastar_generate_protobuf()` to generate its own headers using
ragel, it can benefit from this as well.

so, in change, we specify `SYSTEM` for the exposed target, so that the
compiler does not warn at seeing, for instance, the above switch labels.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>