github: s/ubuntu-latest/ubuntu-24.04/
Ubuntu 24.04 is now available, see
https://github.com/actions/runner-images/issues/9848

so instead of building and running the tests in a fedora 40 container,
let's do the CI with the shinny ubuntu runner image, which preinstalls
some of the necessary building dependencies, so in theory, it should be
relatively faster than installing every package on fedora 40.

and since we are now using an image with git pre-installed, there
is no need to install `git` anymore. so, in this change, we drop the
step to install `git`.

because we are switching to a debian derivative distro, let's use
apt-get instead of dnf to install packages.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>