CMakeLists: Skip CheckGcc107852
as fmt::fmt is introduced to Ceph by including {fmt} project as a subdirectory,
and it is not an alias of a global library.
CheckGcc107852 is not able to find fmt::fmt.
as workaround, comment out the checks and apply the relevant CXXFLAGS.

Co-authored-by: Kefu Chai <tchaikov@gmail.com>
Signed-off-by: Matan Breizman <mbreizma@redhat.com>

Signed-off-by: Matan Breizman <mbreizma@redhat.com>
(cherry picked from commit 275a9103a26ccbaf5a2a8321a040118d1212f156)